% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors_temperature.R
\name{AccDD}
\alias{AccDD}
\title{Accumulated Degree-Days}
\usage{
AccDD(timestamp, temperature, baseline = 10, name = "", plot = F)
}
\arguments{
\item{timestamp}{A vector of timestamps in POSIXct format.}

\item{temperature}{A vector of temperature values.}

\item{baseline}{The baseline temperature for calculating degree days.}

\item{name}{A character string for naming the plot.}

\item{plot}{Logical value indicating whether to plot the delta T values.}
}
\value{
The accumulated degree days (ADD) value.
}
\description{
This function calculates the accumulated degree days (ADD) based
on timed temperature data.
}
