% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{anonymise}
\alias{anonymise}
\title{Anonymise data}
\usage{
anonymise(data, project = NULL, keep_original = FALSE)
}
\arguments{
\item{data}{A data frame to be anonymised.}

\item{project}{A character string with the name of the project.}

\item{keep_original}{A logical value indicating whether to keep the original columns.}
}
\value{
A data frame with the private columns hashed as new columns.
}
\description{
This function anonymises data by hashing private columns defined by helper functions
within the package.
}
