% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_data.R
\name{cs_get_data}
\alias{cs_get_data}
\title{Get data from a community science project}
\usage{
cs_get_data(file = NULL, project = NULL)
}
\arguments{
\item{file}{A character string with the path to the file to read.}

\item{project}{A character string with the name of the project.}
}
\value{
A data frame with the data.
}
\description{
Get data from a community science project
}
