% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries-multi_period.R
\name{daily_delete_missing}
\alias{daily_delete_missing}
\title{Remove days with missing data}
\usage{
daily_delete_missing(data, datecol, max_diff = 30)
}
\arguments{
\item{data}{A data frame containing a date column.}

\item{datecol}{The name of the date column in the data frame.}

\item{max_diff}{The maximum allowed time difference in seconds between
consecutive timestamps. Default is 30 seconds.}
}
\value{
A data frame with the days that have missing data removed.
}
\description{
This function removes days from a dataset where the maximum time difference
between consecutive timestamps exceeds a specified threshold.
}
