% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries-multi_period.R
\name{daily_extremes}
\alias{daily_extremes}
\title{Get the daily extremes from a time series}
\usage{
daily_extremes(values, dates, n = 60, breaks = FALSE)
}
\arguments{
\item{values}{A numeric vector of values.}

\item{dates}{A vector of dates corresponding to the values.}

\item{n}{An integer specifying the number of points to consider for the kernel
min/max calculation. Default is 60.}

\item{breaks}{A logical indicating whether to include the start of the day in the
results. Default is FALSE.}
}
\value{
A data frame with the dates of the daily extremes and their types
("max", "min", and optionally "start" if breaks is TRUE).
}
\description{
This function calculates the daily maximum and minimum values from a time
series of values and their corresponding dates.
}
