% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries-multi_period.R
\name{daily_extremes_stats}
\alias{daily_extremes_stats}
\title{Stats from daily_extremes()}
\usage{
daily_extremes_stats(daily_extremes)
}
\arguments{
\item{daily_extremes}{A data frame returned by \code{daily_extremes()}.}
}
\value{
A data frame with statistics for each day, including minimum, maximum,
range, and their respective timestamps.
}
\description{
Stats from daily_extremes()
}
