% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nature_overheard.R
\name{.natureoverheard_hf_features}
\alias{.natureoverheard_hf_features}
\title{Habitat features in the Nature Overheard dataset}
\usage{
.natureoverheard_hf_features()
}
\value{
A character vector of habitat features
}
\description{
Habitat features in the Nature Overheard dataset
}
