% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flytunes.R
\name{flytunes_annotations}
\alias{flytunes_annotations}
\title{Get sonicscrewdriver annotation objects from FlyTunes data}
\usage{
flytunes_annotations(data)
}
\arguments{
\item{data}{A data frame from \code{flytunes_process_zooniverse()}}
}
\value{
A list of sonicscrewdriver annotation objects
}
\description{
Get sonicscrewdriver annotation objects from FlyTunes data
}
