% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flytunes.R
\name{flytunes_process_zooniverse}
\alias{flytunes_process_zooniverse}
\title{Process FlyTunes export from Zooniverse}
\usage{
flytunes_process_zooniverse(data)
}
\arguments{
\item{data}{A data frame from read.csv of classification csv}
}
\value{
A data frame with processed data
}
\description{
Process FlyTunes export from Zooniverse
}
