% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flytunes.R
\name{flytunes_short_names}
\alias{flytunes_short_names}
\title{Short names for FlyTunes categories}
\usage{
flytunes_short_names()
}
\value{
A character vector of short names
}
\description{
Short names for FlyTunes categories
}
