% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nature_overheard.R
\name{nature_overheard_get_verified_samples}
\alias{nature_overheard_get_verified_samples}
\title{Summarise verifications into samples table}
\usage{
nature_overheard_get_verified_samples(samples, verifications)
}
\arguments{
\item{samples}{A data frame with samples data}

\item{verifications}{A data frame with verifications data}
}
\value{
A data frame: samples data with the verifications data summarised
}
\description{
Summarise verifications into samples table
}
