% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nature_overheard.R
\name{nature_overheard_taxa}
\alias{nature_overheard_taxa}
\title{Get list of taxon column names for Nature Overheard}
\usage{
nature_overheard_taxa()
}
\value{
A character vector of taxon column names
}
\description{
Get list of taxon column names for Nature Overheard
}
