% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors_kernels.R
\name{sensors_kernel_max}
\alias{sensors_kernel_max}
\title{Apply maximum kernel to data}
\usage{
sensors_kernel_max(data, n = 60)
}
\arguments{
\item{data}{A numeric vector of data to be processed.}

\item{n}{An integer specifying the size of the kernel. Default is 60.}
}
\value{
A numeric vector of the same length as the input data.
}
\description{
This function applies a maximum kernel to the data. It uses a
Savitzky-Golay filter to smooth the data and then applies a kernel
to find local maxima.
}
