% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries-single_period.R
\name{sp_extreme}
\alias{sp_extreme}
\title{Find extreme values from a single period}
\usage{
sp_extreme(data, mode = "max")
}
\arguments{
\item{data}{A vector of numeric values representing the data.}

\item{mode}{A character string indicating whether to find the "max" or "min"
extreme value. Default is "max".}
}
\value{
The index of the extreme value within the data vector.
}
\description{
This function finds the extreme value (maximum or minimum) from a single
period of data.
}
