% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries-single_period.R
\name{sp_extremes}
\alias{sp_extremes}
\title{Find extreme values from a single period}
\usage{
sp_extremes(data)
}
\arguments{
\item{data}{A vector of numeric values representing the data.}
}
\value{
The index of the extreme value within the data vector.
}
\description{
This function finds the extreme value (maximum or minimum) from a single
period of data.
}
