% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dates-formats.R
\name{tf_day_seconds}
\alias{tf_day_seconds}
\title{Get the number of seconds into that day from a date object}
\usage{
tf_day_seconds(date)
}
\arguments{
\item{date}{A date object.}
}
\value{
An integer representing the number of seconds into the day.
}
\description{
This function calculates the number of seconds into the day from a date object.
}
