% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dates-formats.R
\name{tf_year_day}
\alias{tf_year_day}
\title{Get a year-day string from a date}
\usage{
tf_year_day(date)
}
\arguments{
\item{date}{A date object.}
}
\value{
A character string in the format "YYYY-DDD".
}
\description{
This function converts a date to a year-day string in the format "YYYY-DDD".
}
