% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifications.R
\name{verifications_get_data}
\alias{verifications_get_data}
\title{Get verifications data from a file}
\usage{
verifications_get_data(file, project = NULL)
}
\arguments{
\item{file}{File path}

\item{project}{Project name}
}
\value{
A data frame with the verifications data
}
\description{
Get verifications data from a file
}
